test: split visual property comment regression
